refactor: Moved EmailValidator to separate namespace
